backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
transformers>=4.37.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0

